+++ src/api/todos.ts
@@ -4,5 +4,3 @@ import { client } from '../utils/fetchClient';
 export const getTodos = (userId: number) => {
   return client.get<Todo[]>(`/todos?userId=${userId}`);
 };
-
-// Add more methods here
