+++ src/components/NewTodo.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export const NewTodo:React.FC = () => (
+  <form>
+    <input
+      type="text"
+      className="todoapp__new-todo"
+      placeholder="What needs to be done?"
+    />
+  </form>
+);
