+++ src/components/TodoItem.tsx
@@ -1,0 +1,70 @@
+import React, { useState } from 'react';
+import cn from 'classnames';
+import { Todo } from '../types/Todo';
+
+interface Props {
+  todo: Todo;
+}
+
+export const TodoItem:React.FC<Props> = ({ todo }) => {
+  const [isEditing, setIsEditing] = useState(false);
+  const [loading] = useState(false);
+
+  return (
+    <div
+      className={cn('todo', {
+        completed: todo.completed,
+      })}
+      key={todo.id}
+    >
+      <label className="todo__status-label">
+        <input
+          type="checkbox"
+          className="todo__status"
+          checked={todo.completed}
+          onChange={() => {}}
+        />
+      </label>
+
+      {isEditing ? (
+        <form>
+          <input
+            type="text"
+            className="todo__title-field"
+            placeholder="Empty todo will be deleted"
+            value="Todo is being edited now"
+          />
+        </form>
+      ) : (
+        <>
+          <span
+            className="todo__title"
+            onDoubleClick={() => setIsEditing(true)}
+          >
+            {todo.title}
+          </span>
+
+          <button
+            type="button"
+            className="todo__remove"
+          >
+            ×
+          </button>
+        </>
+      )}
+
+      {loading && (
+        <div
+          className={cn('modal overlay', {
+            'is-active': loading,
+          })}
+        >
+          <div
+            className="modal-background has-background-white-ter"
+          />
+          <div className="loader" />
+        </div>
+      )}
+    </div>
+  );
+};
