+++ src/index.tsx
@@ -3,7 +3,6 @@ import { createRoot } from 'react-dom/client';
 import 'bulma/css/bulma.css';
 import '@fortawesome/fontawesome-free/css/all.css';
 import './styles/index.scss';
-
 import { App } from './App';
 
 createRoot(document.getElementById('root') as HTMLDivElement)
