+++ src/styles/todoapp.scss
@@ -121,6 +121,7 @@
     appearance: none;
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
+    transition: opacity 0.3s;
 
     &:hover {
       text-decoration: underline;
@@ -128,6 +129,10 @@
 
     &:active {
       text-decoration: none;
+    }
+
+    &:disabled {
+      opacity: 0;
     }
   }
 }
